successful data fetching and displaying

+++ lib/models/movie.dart
@@ -5,7 +5,7 @@ class Movie {
   String posterPath;
   DateTime releaseDate;
   double voteAverage;
-  double durationMinutes;
+  int durationMinutes;
   String ageRestriction;
   List<String> genres;
   List<String> cast;
@@ -25,6 +25,24 @@ class Movie {
       required this.cast,
       required this.tags,
       required this.trendingIndex});
+
+//   @override
+//   String toString() {
+//     return '''
+// Title: $title
+// Backdrop Path: $backDropPath
+// Overview: $overview
+// Poster Path: $posterPath
+// Release Date: $releaseDate
+// Vote Average: $voteAverage
+// Duration Minutes: $durationMinutes
+// Age Restriction: $ageRestriction
+// Genres: ${genres.join(', ')}
+// Cast: ${cast.join(', ')}
+// Tags: ${tags.join(', ')}
+// Trending Index: $trendingIndex
+// ''';
+//   }
 
   // factory Movie.fromJson(Map<String, dynamic> json) {
   //   // Convert genre IDs to corresponding genre names
